Some changes in mutacc/utils/fastq_handler.py
Added function fastq_exclude, which does the same as fastq_extract. i.e.
for a list of read names, a new fastq file is written excluding reads in
the list. This, for large fastq files will mean a lot of 'write'
operations to the disk, to avoid this, a class 'RecordBuffer' is
implemented, writing blocks of fastq_records instead.