Some changes in PR based on discussion on export functionality.
Now, the read names overlapping the region are found with pysam,
and excluded from the input 'background' fastq files with command
line tool seqkit, which I think works well, and fast, for the purpose.

The pedigree structure of the cases is parsed with moonso's ped_parser
using the Individual, and Family classes.

When using mutacc grep now, the user must provide option -m, stating
the family member to make a synthetic fastq file for. if -m 'affected',
A Synthetic fastq based on any affected sample from all of the cases,
regardless of single or trio, will be made.